[ES|QL] Restructure validation code, remove command settings (#215056)
## Summary

Follow on to several recent efforts including
https://github.com/elastic/kibana/issues/195418 and
https://github.com/elastic/kibana/pull/213325

This PR
- reorganizes validation code to make dependencies clearer and make it
less overwhelming... it's not perfect but it's better
- removes the deprecated notion of a command "setting" which only ever
applied to `ENRICH`.

No regression in `ENRICH` mode validation:
<img width="874" alt="Screenshot 2025-03-18 at 1 04 46 PM"
src="https://github.com/user-attachments/assets/e6639d8a-d129-440f-ac30-64a2ef6ab65c"
/>

Or hover
<img width="419" alt="Screenshot 2025-03-18 at 7 43 04 PM"
src="https://github.com/user-attachments/assets/8f9c020c-dcfd-42dc-8e14-4b1c4311457b"
/>

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

---------

(cherry picked from commit 9073b194072cee0ef5290982c05cfdb84662c673)
