+++ src/platform/packages/shared/kbn-esql-validation-autocomplete/index.ts
@@ -12,7 +12,6 @@ export type {
   FunctionDefinition,
   CommandDefinition,
   CommandOptionsDefinition,
-  CommandModeDefinition,
   Literals,
 } from './src/definitions/types';
 export type { ESQLCallbacks } from './src/shared/types';
@@ -51,7 +50,6 @@ export {
   printFunctionSignature,
   checkFunctionArgMatchesDefinition as isEqualType,
   isSourceItem,
-  isSettingItem,
   isFunctionItem,
   isOptionItem,
   isColumnItem,
@@ -61,7 +59,6 @@ export {
   isAssignmentComplete,
   isSingleItem,
 } from './src/shared/helpers';
-export { ENRICH_MODES } from './src/definitions/settings';
 export { timeUnits } from './src/definitions/literals';
 export { aggFunctionDefinitions } from './src/definitions/generated/aggregation_functions';
 export { getFunctionSignatures } from './src/definitions/helpers';
