+++ src/platform/packages/shared/kbn-esql-validation-autocomplete/src/definitions/commands_helpers.ts
@@ -13,6 +13,7 @@ import {
   isFieldExpression,
   Walker,
 } from '@kbn/esql-ast';
+import { i18n } from '@kbn/i18n';
 import {
   getFunctionDefinition,
   isFunctionItem,
@@ -80,3 +81,27 @@ export function checkAggExistence(arg: ESQLFunction): boolean {
 
   return false;
 }
+
+export const ENRICH_MODES = [
+  {
+    name: 'any',
+    description: i18n.translate('kbn-esql-validation-autocomplete.esql.definitions.ccqAnyDoc', {
+      defaultMessage: 'Enrich takes place on any cluster',
+    }),
+  },
+  {
+    name: 'coordinator',
+    description: i18n.translate(
+      'kbn-esql-validation-autocomplete.esql.definitions.ccqCoordinatorDoc',
+      {
+        defaultMessage: 'Enrich takes place on the coordinating cluster receiving an ES|QL',
+      }
+    ),
+  },
+  {
+    name: 'remote',
+    description: i18n.translate('kbn-esql-validation-autocomplete.esql.definitions.ccqRemoteDoc', {
+      defaultMessage: 'Enrich takes place on the cluster hosting the target index.',
+    }),
+  },
+];
