+++ src/platform/packages/shared/kbn-esql-validation-autocomplete/src/validation/esql_validation_meta_tests.json
@@ -9294,7 +9294,7 @@
     {
       "query": "from a_index | enrich _:policy",
       "error": [
-        "Unrecognized value [_] for ENRICH, mode needs to be one of [_ANY, _COORDINATOR, _REMOTE]"
+        "Unrecognized value [_] for ENRICH, mode needs to be one of [_any, _coordinator, _remote]"
       ],
       "warning": []
     },
@@ -9324,7 +9324,7 @@
     {
       "query": "from a_index | enrich any:policy",
       "error": [
-        "Unrecognized value [any] for ENRICH, mode needs to be one of [_ANY, _COORDINATOR, _REMOTE]"
+        "Unrecognized value [any] for ENRICH, mode needs to be one of [_any, _coordinator, _remote]"
       ],
       "warning": []
     },
@@ -9451,7 +9451,7 @@
     {
       "query": "from a_index | enrich _unknown:policy",
       "error": [
-        "Unrecognized value [_unknown] for ENRICH, mode needs to be one of [_ANY, _COORDINATOR, _REMOTE]"
+        "Unrecognized value [_unknown] for ENRICH, mode needs to be one of [_any, _coordinator, _remote]"
       ],
       "warning": []
     },
