+++ src/platform/packages/shared/kbn-monaco/src/esql/lib/hover/hover.test.ts
@@ -11,11 +11,12 @@ import { monaco } from '../../../monaco_imports';
 import { getHoverItem } from './hover';
 import { getAstAndSyntaxErrors } from '@kbn/esql-ast';
 import {
-  ENRICH_MODES,
   ESQLRealField,
   getFunctionDefinition,
   getFunctionSignatures,
 } from '@kbn/esql-validation-autocomplete';
+import { modeDescription } from '@kbn/esql-validation-autocomplete/src/autocomplete/commands/enrich/util';
+import { ENRICH_MODES } from '@kbn/esql-validation-autocomplete/src/definitions/commands_helpers';
 import { FieldType } from '@kbn/esql-validation-autocomplete/src/definitions/types';
 
 const types: FieldType[] = ['keyword', 'double', 'date', 'boolean', 'ip'];
@@ -187,12 +188,11 @@ describe('hover', () => {
     testSuggestions(`from a | enrich policy on b `, 'non-policy', createPolicyContent);
 
     describe('ccq mode', () => {
-      for (const mode of ENRICH_MODES.values) {
-        testSuggestions(
-          `from a | enrich ${ENRICH_MODES.prefix || ''}${mode.name}:policy`,
-          `${ENRICH_MODES.prefix || ''}${mode.name}`,
-          () => [ENRICH_MODES.description, `**${mode.name}**: ${mode.description}`]
-        );
+      for (const mode of ENRICH_MODES) {
+        testSuggestions(`from a | enrich _${mode.name}:policy`, `_${mode.name}`, () => [
+          modeDescription,
+          `**${mode.name}**: ${mode.description}`,
+        ]);
       }
     });
   });
